refactor: moved ReferencedVariable class into new file as public class.
